Move command to index file.
