+++ src/commands/download.ts
@@ -4,4 +4,4 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-export { DownloadCommand as default } from "../download/command";
+export { DownloadCommand as default } from "../download";
